Initial commit
In the first commit of this repository, we add the README.md file into this repository. The file is filled with all the required introductory information that are needed for any user to use the bash-calculator script. More further updates to the documentation will be done in the future, as the script will be updated with more new functions (features).