Added script.sh file
In this commit, we added the script.sh file to this repository. The script.sh file is the main script file for the project "bash-calculator". There are only 4 arithmetic operations that are provided by the script currently. More arithmetic operation features, proper commenting and documentation will be added to the shell script file in the next few days.