Added the feature to calculate factorial, exponentials, and also checking even-odd, prime-composite
In this commit, we have added the feature to calculate a few more arithmetic operations. The new arithmetic operations that we can perform using this script are : Calculate factorial, calculate the exponential value, checking whether the user specified number is even or odd, checking whether the user specified number is prime or composite. The arithmetic operations are done using simple bash calculations and mathematical functions, also using algorithms like looping and calculating. More arithmetic operation features will be added to the script in the coming days. If there are any bugs or errors faced by you or anyone prior to this commit (update /  adding of new features), then kindly apply an issue on this repository.